fix: add-stats: updated StartService - changed the method of obtaining data (filtering at the DataBase level)

+++ src/main/java/ru/yandex/workshop/stats/dto/SellerReportEntry.java
@@ -12,7 +12,8 @@ public class SellerReportEntry {
 
     String productName;
 
-    Integer quantity;
+    Long quantity;
 
-    Float revenue;
+    Double revenue;
+
 }
