+++ src/main/java/ru/yandex/workshop/stats/model/SellerReport.java
@@ -14,5 +14,5 @@ import java.util.List;
 public class SellerReport {
 
     List<SellerReportEntry> sellerReportEntryList;
-    Float sumRevenue;
+    Double sumRevenue;
 }
